[Feat] - 내정보조회, 사용자 회원가입,  로그인 API(CRS 코드분리)_01
