+++ eslint.config.js
@@ -1,0 +1,8 @@
+import globals from "globals";
+import pluginJs from "@eslint/js";
+
+
+export default [
+  {languageOptions: { globals: globals.node }},
+  pluginJs.configs.recommended,
+];
