+++ package.json
@@ -19,6 +19,9 @@
     "prisma": "^5.14.0"
   },
   "devDependencies": {
+    "@eslint/js": "^9.4.0",
+    "eslint": "9.x",
+    "globals": "^15.4.0",
     "nodemon": "^3.1.0",
     "prettier": "^3.2.5"
   }
